Refactor: DatoVelger som en selvstandig komponent
